refactor && CodeGen abstraction && IR analysis bug fix
