+++ ort_aot/common.py
@@ -1,9 +1,129 @@
 from collections import defaultdict, OrderedDict
 import onnx
 import copy
+import enum
+import numpy as np
 import symbolic_shape_infer
+from abc import ABCMeta, abstractmethod
+import logger
+
+_ENABLE_PARALLEL_COMPILE = False
+
+class DataType(enum.Enum):
+    UNDEFINED = 0
+    # Basic types.
+    FLOAT = 1   # float
+    UINT8 = 2   # uint8_t
+    INT8 = 3    # int8_t
+    UINT16 = 4  # uint16_t
+    INT16 = 5   # int16_t
+    INT32 = 6   # int32_t
+    INT64 = 7   # int64_t
+    STRING = 8  # string
+    BOOL = 9    # bool
+    # Advanced types
+    FLOAT16 = 10
+    DOUBLE = 11
+    UINT32 = 12
+    UINT64 = 13
+    COMPLEX64 = 14     # complex with float32 real and imaginary components
+    COMPLEX128 = 15    # complex with float64 real and imaginary components
+    # Future extensions go here.
 
 
+class AttributeType(enum.Enum):
+    UNDEFINED = 0
+    FLOAT = 1
+    INT = 2
+    STRING = 3
+    TENSOR = 4
+    GRAPH = 5
+    SPARSE_TENSOR = 11
+    TYPE_PROTO = 13
+
+    FLOATS = 6
+    INTS = 7
+    STRINGS = 8
+    TENSORS = 9
+    GRAPHS = 10
+    SPARSE_TENSORS = 12
+    TYPE_PROTOS = 14
+
+
+# This map is used for converting TensorProto values into Numpy arrays
+TENSOR_TYPE_TO_NP_TYPE = {
+    int(onnx.TensorProto.FLOAT): np.dtype('float32'),
+    int(onnx.TensorProto.UINT8): np.dtype('uint8'),
+    int(onnx.TensorProto.INT8): np.dtype('int8'),
+    int(onnx.TensorProto.UINT16): np.dtype('uint16'),
+    int(onnx.TensorProto.INT16): np.dtype('int16'),
+    int(onnx.TensorProto.INT32): np.dtype('int32'),
+    int(onnx.TensorProto.INT64): np.dtype('int64'),
+    int(onnx.TensorProto.BOOL): np.dtype('bool'),
+    int(onnx.TensorProto.FLOAT16): np.dtype('float16'),
+    # Native numpy does not support bfloat16 so now use float32 for bf16 values
+    int(onnx.TensorProto.BFLOAT16): np.dtype('float32'),
+    int(onnx.TensorProto.DOUBLE): np.dtype('float64'),
+    int(onnx.TensorProto.COMPLEX64): np.dtype('complex64'),
+    int(onnx.TensorProto.COMPLEX128): np.dtype('complex128'),
+    int(onnx.TensorProto.UINT32): np.dtype('uint32'),
+    int(onnx.TensorProto.UINT64): np.dtype('uint64'),
+    int(onnx.TensorProto.STRING): np.dtype('object')
+}
+
+# Currently native numpy does not support bfloat16 so TensorProto.BFLOAT16 is ignored for now
+# Numpy float32 array is only reversed to TensorProto.FLOAT
+NP_TYPE_TO_TENSOR_TYPE = {
+    v: k for k, v in TENSOR_TYPE_TO_NP_TYPE.items() if k != onnx.TensorProto.BFLOAT16}
+
+import numpy
+NP_TYPE_C_TYPE = {numpy.bool_: 'bool',
+ numpy.byte:
+ 'int8_t',
+ numpy.ubyte:
+ 'uint8_t',
+ numpy.short:
+ 'short',
+ numpy.ushort:
+ 'unsigned short',
+ numpy.intc:
+ 'int',
+ numpy.uintc:
+ 'unsigned int',
+ numpy.int_:
+ 'long',
+ numpy.uint:
+ 'unsigned long',
+ numpy.longlong:
+ 'long long',
+ numpy.ulonglong:
+ 'unsigned long long',
+ numpy.single:
+ 'float',
+ numpy.float32:'float',
+ numpy.double:
+ 'double',
+ numpy.longdouble:
+ 'long double',
+ }
+
+
+
+class NodeVisitor(object):
+    def __init__(self):
+        pass
+    
+    #interface for codegen/lowering
+    @abstractmethod
+    def visit(self, node , context, indent: int):
+        pass
+
+class CodeGenContext(object):
+    def __init__(self, var_map: dict):
+        self.var_map = var_map
+        self.vectorized_var_set: Dict = set()
+
+ 
 class SpecialVar(object):
     def __init__(self):
         self.input_args = "input_args"
@@ -11,6 +131,7 @@ class SpecialVar(object):
         self.input_args_size = "input_args_size"
         self.parallel_loop_start = "p_loop_start"
         self.parallel_loop_end = "p_loop_end"
+        self.dynamic_shape_args = "dynamic_shape_args"
 
 
 def add_all_intermidiate_values(model):
@@ -20,7 +141,8 @@ def add_all_intermidiate_values(model):
     for node in model_proto.graph.node:
         for output in node.output:
             if output not in org_outputs:
-                model_proto.graph.output.extend([onnx.ValueInfoProto(name=output)])
+                model_proto.graph.output.extend(
+                    [onnx.ValueInfoProto(name=output)])
     return model_proto
 
 
@@ -29,11 +151,36 @@ def get_symbol_shape(model_path):
         model = onnx.load(model_path)
     else:
         model = model_path
+    symbolic_shape_infer.logger.setLevel(symbolic_shape_infer.logging.ERROR)
     symbol_shape = symbolic_shape_infer.SymbolicShapeInference.infer_shapes(
         model, 2**31 - 1, True, guess_output_rank=True, verbose=1
     )
 
     return symbol_shape
+
+
+def parse_onnx_to_numpyarray(value):    
+    data = None
+    if isinstance(value, onnx.onnx_ml_pb2.NodeProto):
+        assert len(value.attribute) == 1
+        Attr_type = AttributeType(value.attribute[0].type)
+        if Attr_type == AttributeType.INT:
+            data = np.array([value.attribute[0].i])
+        elif Attr_type == AttributeType.TENSOR:
+            value = value.attribute[0].t
+        elif Attr_type == AttributeType.FLOAT:
+            data = np.array([value.attribute[0].f])
+        else:
+            raise Exception("not support")
+    # handle      AttributeType.TENSOR
+    if isinstance(value, onnx.onnx_ml_pb2.TensorProto):
+        assert data is None
+        data = onnx.numpy_helper.to_array(value)
+    elif data is not None:
+        pass
+    else:
+        assert RuntimeError("not support proto type")
+    return data
 
 
 class OnnxInGraph(object):
@@ -85,7 +232,8 @@ class OnnxInGraph(object):
         inputs = tokenizer("Hello, my dog is cute", return_tensors="np")
         # inputs["input_mask"] = inputs["attention_mask"]
         # del inputs["attention_mask"]
-        sess = ort.InferenceSession(path_or_bytes=model_proto.SerializeToString())
+        sess = ort.InferenceSession(
+            path_or_bytes=model_proto.SerializeToString())
         if sess.get_inputs()[0].name == "input_ids":
             ret = sess.run([i.name for i in sess.get_outputs()], dict(inputs))
             runtime_shape = {
@@ -133,8 +281,10 @@ class OnnxInGraph(object):
             self.node_name2module[
                 "out_" + out.name
             ] = out  # add `out_` in case the output has the same name with the last node
-        self.graph_output_names = ["out_" + out.name for out in self.graph.output]
-        symbol_shape, rt_shape = self.get_all_shape_from_onnx_model(self.model_proto)
+        self.graph_output_names = [
+            "out_" + out.name for out in self.graph.output]
+        symbol_shape, rt_shape = self.get_all_shape_from_onnx_model(
+            self.model_proto)
         self.tensor_type_shape_info.update(symbol_shape)
         self.rt_shape = rt_shape
 
@@ -161,6 +311,7 @@ class IndexSubGraph(object):
                 self.input_name_ref_c.pop(ink)
                 if len(consumed_by_name2node[ink]) == v:
                     self.output_name_ref_c.pop(ink)
+
 
 class HardwareContext(object):
     def __init__(self, device_id, vec_lanes):
