+++ ort_aot/sympy_utils.py
@@ -1,4 +1,5 @@
 import sympy
+import re
 from sympy.core.logic import fuzzy_and, fuzzy_or  # type: ignore[import]
 
 
@@ -10,7 +11,14 @@ def sympy_symbol(name):
         return sympy.Integer(name)
     elif isinstance(name, list):
         return [sympy_symbol(x) for x in name]
+    elif isinstance(name, str):
+        if ' ' in name or ',' in name:
+            name = re.sub(r'[^a-zA-Z0-9_]+', '_', name)
     return sympy.Symbol(name, integer=True, positive=True)
+
+def sympy_dot(seq1, seq2):
+    assert len(seq1) == len(seq2)
+    return sympy.expand(sum(a * b for a, b in zip(seq1, seq2)))
 
 class FloorDiv(sympy.Function):
         """
